requirements.txt
fastapi
uvicorn
circuitpython-build-tools
intelhex
python-dotenv
python-multipart
